chatbot_itributos/requirements.txt
streamlit==1.39.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
diskcache==5.6.3
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
